[#169211] Set an overriden Jackson Jaxb provider to avoid use of Johzon
See Stackoverflow : How to use Jackson as JSON provider for
JAX-RS-Client instead of Johnzon in TomEE 7?
https://stackoverflow.com/a/38915278

quote : « jackson uses a custom matching strategy and therefore uses
"/". So the bus priority is ignored since mediatype priority is enough
to say johnzon is "more adapted" than jackson for json.
 * To solve it the easiest is likely to override jackson provider (just
extend it) and decorate it with @Provides/@Consumes
withediaType.APPLICATION_JSON instead of wildcard one »